requirements.txt
django
python-dotenv
mysqlclient
djangorestframework
faker
mysql
